backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]

